feat(reporting): wrapper/checker/emailer for reporting
This includes a simple wrapper/checker/emailer
for the reporting script.

Scope: Since DROP commands are in all reporting queries,
if the query fail, the table may be missing afterwards. This
commit include functions to ensure completion of the
entire report.

A pre-defined list of views is listed in `list_of_views_sorted`,
which is verified by `check_report_tables.sh` by using
a new SQL query call - checking_completion.sql.

A wrapper ensure try 5 times to create the tables
with reporting_script.sh, store the logs properly,
and send an email to the respective persons.

sendmail + ssmtp need to be configured for
email to work.